Create settings submodule.
